Apply dynamic imports for useTimesSquarePage users
This comprehensive change applies the dynamic import pattern to all
remaining components that use useTimesSquarePage hook, preventing SSR
execution of SWR:

**Components Updated:**

1. **TimesSquareGitHubPagePanel** → TimesSquareGitHubPagePanelClient
   - Sidebar panel with page metadata and controls
   - Loading fallback during SSR

2. **TimesSquareParameters** → TimesSquareParametersClient
   - Interactive parameter form with JSON schema validation
   - Client-only rendering to prevent form hydration issues

3. **TimesSquareHtmlEventsProvider** → TimesSquareHtmlEventsProviderClient
   - SSE context provider for real-time execution updates
   - Default context values provided during SSR loading
   - Prevents Server-Sent Events from running on server

**Technical Details:**
- All components use dynamic imports with ssr: false
- Appropriate loading states and fallbacks for each component type
- Context providers maintain API compatibility with default values
- Prevents useTimesSquarePage SWR hook from executing during SSR

This completes the SSR compatibility work for all Times Square SWR hooks
while maintaining full client-side functionality.
